refactor(etf): dedupe holdings column logic and align theme
- Extract shared buildEtfHoldingColumnDefs + field maps into
  utils/etf-holdings-utils.ts; EtfHoldings and EtfMorInfo now import
  the same helper instead of each maintaining their own copy.
- Swap the hardcoded indigo "View more holdings" link color for the
  site's themed link-color class so it tracks the global accent.
- Let EtfHoldings derive a contextual title ("Top N Holdings" when
  previewing a truncated list, "Holdings" otherwise) instead of the
  ambiguous static default.
